cloud-run/trickle-in/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
numerapi==2.9.4
pandas==1.3.2
google-cloud-storage==1.18.1
google-cloud-bigquery==1.21.0
google-cloud-pubsub==2.8.0
pyarrow==3.0.0
